{"schema":"decodex/commit/1","summary":"{\"schema\":\"decodex/commit/1\",\"summary\":\"Split live adapter domain materializers\",\"authority\":\"manual\"}","authority":"manual"}
